prerefactoring

+++ DotNet2020/DotNet2020.Domain.3/Models/SpecificWorkerCompetencesModel.cs
@@ -7,7 +7,7 @@ namespace DotNet2020.Domain._3.Models
 {
     public class SpecificWorkerCompetencesModel
     {
-        public long WorkerId { get; set; }
+        public int WorkerId { get; set; }
         public SpecificWorkerModel Worker { get; set; }
         public long CompetenceId { get; set; }
         public CompetencesModel Competence { get; set; }
